remove old map files
